users-service/requirements.txt
grpcio==1.28.0.dev0
grpcio-tools==1.28.0.dev0
pymongo
dnspython==1.16.0

